requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
coverage==6.3.2
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
greenlet==1.1.2
iniconfig==1.1.1
jsonschema==4.4.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==1.7.1
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.0.1
pytest-cov==3.0.0
pytest-django==4.5.2
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
tzdata==2021.5
